requirements.txt
amqp==2.6.0
analytical==0.0.1
asgiref==3.2.7
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.5
certifi==2020.4.5.2
cffi==1.14.0
channels==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.42.0
django-cors-headers==3.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
future==0.18.2
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.1
incremental==17.5.0
kombu==4.6.10
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2020.1
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
vine==1.3.0
whitenoise==5.1.0
zipp==3.1.0
zope.interface==5.1.0
